--- src/main/java/me/drison64/mypin/Configurations/Configuration.java
@@ -1,83 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ondřej Vajďák
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package me.drison64.mypin.Configurations;
-
-import me.drison64.mypin.Main;
-import me.drison64.mypin.Managers.ConfigManager;
-import org.bukkit.Bukkit;
-import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.configuration.file.YamlConfiguration;
-
-import java.io.File;
-import java.io.IOException;
-
-public class Configuration {
-
-    private Main main;
-    private ConfigManager configManager;
-
-    public Configuration(Main main) {
-        this.main = main;
-        this.configManager = main.getConfigManager();
-    }
-
-    protected File file;
-    protected FileConfiguration config;
-    protected String fileName;
-
-    public void createConfig(String fileName) {
-        this.fileName = fileName;
-        file = new File(Bukkit.getPluginManager().getPlugin("MyPin").getDataFolder(), fileName + ".yml");
-        if (!file.exists()) {
-            try {
-                file.createNewFile();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-        config = YamlConfiguration.loadConfiguration(file);
-    }
-
-    public void reloadConfig() {
-        config = YamlConfiguration.loadConfiguration(file);
-    }
-
-    public FileConfiguration get() {
-        return config;
-    }
-
-    public void save() {
-        try {
-            config.save(file);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public String getFileName() {
-        return fileName;
-    }
-
-}
