+++ src/main/java/me/drison64/mypin/Main.java
@@ -26,17 +26,17 @@ package me.drison64.mypin;
 
 import me.drison64.inventoryapi.InventoryAPI;
 import me.drison64.inventoryapi.InventoryManager;
-import me.drison64.mypin.Configurations.Config;
-import me.drison64.mypin.Configurations.Data;
-import me.drison64.mypin.Listener.PlayerListener;
-import me.drison64.mypin.Managers.ActionsManager;
-import me.drison64.mypin.Managers.ConfigManager;
-import me.drison64.mypin.Managers.PinManager;
-import me.drison64.mypin.Managers.WaitingManager;
-import me.drison64.mypin.Objects.Action.*;
-import me.drison64.mypin.Utils.DefaultActionsUtils;
-import me.drison64.mypin.Utils.DoorUtils;
-import me.drison64.mypin.Utils.PinUtils;
+import me.drison64.mypin.configurations.Config;
+import me.drison64.mypin.configurations.Data;
+import me.drison64.mypin.listener.PlayerListener;
+import me.drison64.mypin.managers.ActionsManager;
+import me.drison64.mypin.managers.ConfigManager;
+import me.drison64.mypin.managers.PinManager;
+import me.drison64.mypin.managers.WaitingManager;
+import me.drison64.mypin.objects.Action.*;
+import me.drison64.mypin.utils.DefaultActionsUtils;
+import me.drison64.mypin.utils.DoorUtils;
+import me.drison64.mypin.utils.PinUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
