--- src/main/java/me/drison64/mypin/Objects/Action/Action.java
@@ -1,59 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ondřej Vajďák
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package me.drison64.mypin.Objects.Action;
-
-import me.drison64.mypin.Main;
-import org.bukkit.block.Block;
-import org.bukkit.entity.Player;
-
-import java.util.List;
-
-public abstract class Action {
-
-    private Main main;
-    private ActionType type;
-
-    public Action(Main main, ActionType type) {
-        this.main = main;
-        this.type = type;
-    }
-
-    public abstract void run(List<String> input, Integer line, Block block, Player player);
-
-    public ActionType getType() {
-        return type;
-    }
-
-    protected void runNext(List<String> input, Integer line, Block block, Player player) {
-
-        if (input.size() < line + 1) {
-            return;
-        }
-
-        main.getActionsManager().get(type).run(input, line + 1 , block, player);
-
-    }
-
-}
