--- src/main/java/me/drison64/mypin/Objects/Action/Action_PlayerCommand.java
@@ -1,64 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ondřej Vajďák
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package me.drison64.mypin.Objects.Action;
-
-import me.drison64.mypin.Main;
-import me.drison64.mypin.Utils.StringStitcherUtils;
-import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.block.Block;
-import org.bukkit.entity.Player;
-
-import java.util.List;
-
-public class Action_PlayerCommand extends Action {
-
-    private Main main;
-    private String[] splitted;
-    private String command;
-    private int delay = 1;
-
-    public Action_PlayerCommand(Main main, ActionType type) {
-        super(main, type);
-        this.main = main;
-    }
-
-    @Override
-    public void run(List<String> data, Integer line, Block block, Player player) {
-
-        this.splitted = data.get(line - 1).split(" ");
-
-        if (splitted.length < 2) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.DARK_RED + "Error occured at line: " + (line + 1) + ", value is empty.");
-        }
-
-        command = StringStitcherUtils.stitch(splitted, 1);
-
-        player.performCommand(command);
-        runNext(data, line, block, player);
-
-    }
-
-}
