--- src/main/java/me/drison64/mypin/Objects/ConfigType.java
@@ -1,45 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ondřej Vajďák
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package me.drison64.mypin.Objects;
-
-import me.drison64.mypin.Configurations.Config;
-import me.drison64.mypin.Configurations.Configuration;
-import me.drison64.mypin.Configurations.Data;
-
-public enum ConfigType {
-
-    CONFIG(Config.class),
-    DATA(Data.class);
-
-    private Class<? extends Configuration> clazz;
-
-    ConfigType(Class<? extends Configuration> clazz) {
-        this.clazz = clazz;
-    }
-
-    public Class<? extends Configuration> getClazz() {
-        return clazz;
-    }
-}
