--- src/main/java/me/drison64/mypin/Utils/EncryptionUtils.java
@@ -1,90 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ondřej Vajďák
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package me.drison64.mypin.Utils;
-
-import me.drison64.mypin.Main;
-import org.bukkit.Location;
-
-import java.math.BigInteger;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class EncryptionUtils {
-
-    private Main main;
-
-    public EncryptionUtils(Main main) {
-        this.main = main;
-    }
-
-    public static String toSHA256(String input) {
-        String hash;
-
-        //HASH
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            byte[] messageDigest = md.digest(input.getBytes());
-            BigInteger no = new BigInteger(1, messageDigest);
-            String hashtext = no.toString(16);
-            while (hashtext.length() < 32) {
-                hashtext = "0" + hashtext;
-            }
-            hash = hashtext;
-        } catch (NoSuchAlgorithmException e) {
-            System.out.println("Exception thrown" + " for incorrect algorithm: " + e);
-            return null;
-        }
-
-        return hash;
-    }
-
-    public static String toSHA256(String input, Location loc) {
-        String hash;
-
-        //HASH
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            byte[] messageDigest = md.digest(input.getBytes());
-            BigInteger no = new BigInteger(1, messageDigest);
-            String hashtext = no.toString(16);
-            while (hashtext.length() < 32) {
-                hashtext = "0" + hashtext;
-            }
-            hash = hashtext;
-        } catch (NoSuchAlgorithmException e) {
-            System.out.println("Exception thrown" + " for incorrect algorithm: " + e);
-            return null;
-        }
-
-        String slocx = String.valueOf(loc.getBlockX());
-        String slocy = String.valueOf(loc.getBlockY());
-        String slocz = String.valueOf(loc.getBlockZ());
-        String sloc = slocx + slocy + slocz;
-
-        return hash + sloc;
-    }
-
-
-}
