--- src/main/java/me/drison64/mypin/Utils/StringStitcherUtils.java
@@ -1,45 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Ondřej Vajďák
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package me.drison64.mypin.Utils;
-
-public class StringStitcherUtils {
-
-    public static String stitch(String[] input) {
-        return doStitch(input, 0);
-    }
-
-    public static String stitch(String[] input, Integer startIndex) {
-        return doStitch(input, startIndex);
-    }
-
-    private static String doStitch(String[] input, Integer startIndex) {
-        String output = "";
-        for (int i = startIndex; i < input.length; i++) {
-            output = output + input[i];
-        }
-        return output.substring(1, output.length());
-    }
-
-}
